% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_cardpoints.R
\name{add_foreign_countries}
\alias{add_foreign_countries}
\title{Add columns with statistics on transactions in foreign countries}
\usage{
add_foreign_countries(sf_cardpoints)
}
\arguments{
\item{sf_cardpoints}{A geometry object of class sf containing points at which internet transactions took place.}
}
\value{
sf_cardpoints with additional columns.
}
\description{
Add columns with statistics on transactions in foreign countries
}
