% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mobility_results.R
\name{add_mobility_results}
\alias{add_mobility_results}
\title{Add mobility results}
\usage{
add_mobility_results(features, sf_cardpoints)
}
\arguments{
\item{features}{A data frame to which new features will be added.}

\item{sf_cardpoints}{geocoded transaction data stored in spatial data frame of class sf with geometry "POINT".}
}
\value{
The features data frame with additional columns.
}
\description{
Add mobility results
}
