% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_raster_features.R
\name{add_raster_features}
\alias{add_raster_features}
\title{Aggregate and add raster features}
\usage{
add_raster_features(features, df_sp, RASTERS, RASTERSTACK_PATH)
}
\arguments{
\item{features}{A data frame to which new features will be added.}

\item{df_sp}{A spatial data frame of class sf with geometry "POINT", for which raster values will be extracted - and aggregated by its id.}

\item{RASTERS}{Character vector. Names of raster files (without file-extension).}

\item{RASTER_FILTERED}{String. Path to directory containing raster files.}
}
\value{
The features data frame with additional columns.
}
\description{
Aggregate and add raster features
}
