% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spending_results.R
\name{add_spending_results}
\alias{add_spending_results}
\title{Add spending results}
\usage{
add_spending_results(features, df_raw, GOOGLE_COLUMN, AMOUNT_COLUMN,
  CUSTOMER_ID_COLUMN)
}
\arguments{
\item{features}{A data frame to which new features will be added.}

\item{df_raw}{A data frame containing transaction data.}

\item{GOOGLE_COLUMN}{String. Name of the column of df containing business type (from Google API).}

\item{AMOUNT_COLUMN}{String. Name of the column of df containing the amount.}

\item{CUSTOMER_ID_COLUMN}{String. Name of the column of df containing the customer identifier.}
}
\value{
The features data frame with additional columns.
}
\description{
Add spending results
}
