% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_features.R
\name{add_time_features}
\alias{add_time_features}
\title{Add time features to a feature collection}
\usage{
add_time_features(features, df, CUSTOMER_ID_COLUMN, TIME_COLUMN)
}
\arguments{
\item{features}{A data frame to which new features will be added.}

\item{df}{A data frame containing the columns specified by the following arguments.}

\item{CUSTOMER_ID_COLUMN}{String. Name of the column of df containing the customer identifier.}

\item{TIME_COLUMN}{String. Name of the column of df containing timestamps.}
}
\value{
The features data frame with additional columns.
}
\description{
Add time features to a feature collection
}
