% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{apply_mappings}
\alias{apply_mappings}
\title{Add features by mapping certain columns to business type and price type}
\usage{
apply_mappings(df, BuchT_Mapping_json, GoogT_Mapping_json,
  BusiT_Mapping_json, Price_Mapping_json, LOCATION_COLUMN,
  ALT_LOCATION_COLUMN, BOOKING_COLUMN, GOOGLE_COLUMN)
}
\arguments{
\item{df}{Dataframe with transaction data.}

\item{BuchT_Mapping_json}{Mapping for booking categories on target categories.}

\item{GoogT_Mapping_json}{Mapping for google categories on target categories.}

\item{BusiT_Mapping_json}{Mapping for booking text strings  on target categories.}

\item{Price_Mapping_json}{Mapping for booking texts strings on price categories.}

\item{LOCATION_COLUMN}{Column storing booking text strings.}

\item{ALT_LOCATION_COLUMN}{Column storing alternative text strings.}

\item{BOOKING_COLUMN}{Column storing booking categories text strings.}

\item{GOOGLE_COLUMN}{Column storing google categories text strings.}
}
\value{
The data frame df with additional columns (target categories) 'businesstype' and 'pricestype'.
}
\description{
Add features by mapping certain columns to business type and price type
}
