% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet_points.R
\name{create_internetpoints}
\alias{create_internetpoints}
\title{Extract all transactions that likely took place via the internet}
\usage{
create_internetpoints(df_raw, LAT_COLUMN, LON_COLUMN)
}
\arguments{
\item{df_raw}{Data frame containing coordinate and businesstype columns.}

\item{LAT_COLUMN}{String. Name of the column of df containing the latitude of location.}

\item{LON_COLUMN}{String. Name of the column of df containing the longitude of location.}
}
\value{
A geometry object (class sf) containing points at which internet transactions took place.
}
\description{
Extract all transactions that likely took place via the internet
}
