% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{.default_if_file_not_exists}
\alias{.default_if_file_not_exists}
\title{Full path of file in folder}
\usage{
.default_if_file_not_exists(inputs_path, file_name)
}
\arguments{
\item{inputs_path}{The wanted directory.}

\item{file_name}{The name of the wanted file.}
}
\value{
A string representing the path of the file.
}
\description{
Returns the full file path of a file in a given directory, if it is located there.
Else the path is relative to the sGEOkb package inst-folder.
}
\details{
This function is used to load configuration from a file in a directory that is
passed as a command line argument, or - if the file does not exist in that directory -
to load the default configuration file shipped with the sGEOkb package.
}
