% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_features.R
\name{evaluate_today_freeday}
\alias{evaluate_today_freeday}
\title{Check if a given day was free of work}
\usage{
evaluate_today_freeday(dt)
}
\arguments{
\item{dt}{Chracter vector. The date time.}
}
\value{
A boolean character of the same length as dt.
}
\description{
Check if a given day was free of work
}
