% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_default_values.R
\name{fill_default_values}
\alias{fill_default_values}
\title{Replace missing values with defaults where applicable}
\usage{
fill_default_values(features, MEDIANthreshold = 0.05)
}
\arguments{
\item{features}{Data frame with missing values.}

\item{MEDIANthreshold}{Threshold in percent: If the number of missings abount to a share of less than this value, NAs will be replaces with median. Default is 0.05.}
}
\value{
The feature data frame where missing values have been replaced.
}
\description{
Replace missing values with defaults where applicable
}
