% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chdf_buffer.R
\name{get_chdf_buffer}
\alias{get_chdf_buffer}
\title{Get a buffer object for Switzerland}
\usage{
get_chdf_buffer(gdf, SWISS_BUFFER)
}
\arguments{
\item{gdf}{A spatial data frame of class sf with geometry "POLYGON" representing the countries of the world.}

\item{SWISS_BUFFER}{Numeric. Distance in meters to Swiss border, under which the area is interpreted as close-to Swizerland.}
}
\value{
A named list with two elements (chdf_buffer and chdf_3035 sf objects objects with geometry "POLYGON" representing Swizerland and its proximity, enlarged by a buffer).
}
\description{
Get a buffer object for Switzerland
}
