% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_features.R
\name{get_features}
\alias{get_features}
\title{Extract time features}
\usage{
get_features(customer_df, customer_id)
}
\arguments{
\item{customer_df}{The transaction data frame containing only records associated with a single customer.}

\item{customer_id}{THe ID of the single customer associated with customer_df.}
}
\value{
customer_df with additional columns.
}
\description{
Extract time features
}
