% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_raster_features.R
\name{get_rasterstack_features}
\alias{get_rasterstack_features}
\title{Extract raster values from given coordinates}
\usage{
get_rasterstack_features(gpd_coords_spdf, RASTERSTACK_PATH, RASTERS)
}
\arguments{
\item{gpd_coords_spdf}{A spatial data frame of class sf with geometry "POINT", for which raster values will be extracted.}

\item{RASTERSTACK_PATH}{String. Path to directory containing raster files.}

\item{RASTERS}{Character vector. Names of raster files (without file-extension) that should be used.}
}
\value{
gpd_coords_spdf with additional columsn representing features from the raster files.
}
\description{
Extract raster values from given coordinates
}
