% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_cardpoints.R
\name{get_sf_cardpoints}
\alias{get_sf_cardpoints}
\title{Create sf object of card transaction data and compute geographic/mobility related information}
\usage{
get_sf_cardpoints(df_sp, gdf, internetpoints, chdf_buffer, chdf_3035,
  SWISS_BUFFER)
}
\arguments{
\item{df_sp}{A spatial data frame of class sf with geometry "POINT" representing transaction data.}

\item{gdf}{A spatial data frame of class sf with geometry "POLYGON" representing the countries of the world.}

\item{internetpoints}{A geometry object of class sf containing points at which internet transactions took place.}

\item{chdf_buffer}{sf objects objects with geometry "POLYGON" representing Swizerland and its proximity, enlarged by a buffer.}

\item{chdf_3035}{sf objects objects with geometry "POLYGON" in crs with unit=meter representing Swizerland and its proximity, enlarged by a buffer.}

\item{SWISS_BUFFER}{Numeric. Distance in meters to Swiss border, under which the area is interpreted as close-to Swizerland.}
}
\value{
The sdf_sp data frame with additional columns.
}
\description{
Add some features related to distance (centroids per user) and area (space covered by user transactions).
}
