% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{list_required_variables}
\alias{list_required_variables}
\title{List all required variables of a parameters file}
\usage{
list_required_variables()
}
\value{
A character vector containing the variable names.
}
\description{
List all required variables of a parameters file
}
