% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{load_mapping_files}
\alias{load_mapping_files}
\title{Load mapping files}
\usage{
load_mapping_files(inputs_path)
}
\arguments{
\item{inputs_path}{String. Path to the directory containing the mapping files.}
}
\value{
A list of mapping files
}
\description{
Load mapping files
}
