% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Load Parameters from Parameter File}
\usage{
load_parameters(file = NULL)
}
\arguments{
\item{file}{String. The path to the yaml-document containing the parameter definitions.}
}
\value{
A named list containing the parameters as key-value pairs.
}
\description{
Load Parameters from Parameter File
}
