% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Compute features for input data}
\usage{
main(args = NULL, dev_mode = FALSE)
}
\arguments{
\item{args}{A named list containing the arguments as key-value pairs.}

\item{dev_mode}{Boolean. Run this function in development mode with test data
and verbose logging?}
}
\value{
A boolean indicating successful computation.
This function is called for its side effects. Results are written to a data base.
}
\description{
Compute features for input data
}
