% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_format.R
\name{transform_to_spatial_format}
\alias{transform_to_spatial_format}
\title{Transform data frame to spatial format}
\usage{
transform_to_spatial_format(df, LAT_COLUMN, LON_COLUMN, TIME_COLUMN,
  CUSTOMER_ID_COLUMN)
}
\arguments{
\item{df}{A data frame containing the columns specified by the following arguments.}

\item{LAT_COLUMN}{String. Name of the column of df containing latitude values.}

\item{LON_COLUMN}{String. Name of the column of df containing longtitude values.}

\item{TIME_COLUMN}{String. Name of the column of df containing timestamps.}

\item{CUSTOMER_ID_COLUMN}{String. Name of the column of df containing the customer identifier.}
}
\value{
A spatial data frame of class sf
}
\description{
This functions takes a data frame and extracts location information
(longitude and latitude columns) as well as a customer identifier
into a spatial data frame of class sf ('simple features collection'),
orderd by customer id and time.
}
