% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_connection_config}
\alias{validate_connection_config}
\title{Validate connection configuration file}
\usage{
validate_connection_config(connection_config)
}
\arguments{
\item{connection_config}{The path to the config file.}
}
\value{
A boolean.
}
\description{
Yaml file should have one element 'kbda_lab' with several children.
}
