% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_countries_data}
\alias{validate_countries_data}
\title{Validate countries data}
\usage{
validate_countries_data(countries)
}
\arguments{
\item{countries}{A data frame of class sf.}
}
\value{
A boolean indicating if countries is valid.
}
\description{
Validate countries data
}
