% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_mapping_files}
\alias{validate_mapping_files}
\title{Validate mapping files}
\usage{
validate_mapping_files(mapping_files)
}
\arguments{
\item{mapping_files}{A list of mapping files.}
}
\value{
A boolean.
}
\description{
Validate mapping files
}
