% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Check that required parameters are present}
\usage{
validate_parameters(params)
}
\arguments{
\item{params}{A named list containing parameters as key-value pairs.}
}
\value{
A boolean indicating if the validation was successful.
}
\description{
Check that required parameters are present
}
