% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_raster_files}
\alias{validate_raster_files}
\title{Validate raster files}
\usage{
validate_raster_files(RASTERSTACK_PATH, RASTERS)
}
\arguments{
\item{RASTERSTACK_PATH}{String. The path to the directory containing the raster files.}

\item{RASTERS}{Character vector containing the names of the raster files.}
}
\value{
A boolean.
}
\description{
Validate raster files
}
