% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_results}
\alias{validate_results}
\title{Validate result data}
\usage{
validate_results(features, df, CUSTOMER_ID_COLUMN, MISSING_THRESHOLD_PCT)
}
\arguments{
\item{features}{The data frame containing the result data}
}
\value{
A boolean
}
\description{
Validate result data
}
