% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_transaction_data}
\alias{validate_transaction_data}
\title{Validate transaction data}
\usage{
validate_transaction_data(transaction_data = NULL, params = NULL)
}
\arguments{
\item{transaction_data}{A data frame.}

\item{params}{A named list of parameters such as returned by load_parameters().}
}
\value{
A boolean.
}
\description{
Validate transaction data
}
